+++ feature-gamescreen/src/main/java/com/fiz/battleinthespace/feature_gamescreen/data/actor/weapon/Ball.kt
@@ -45,7 +45,7 @@ class Ball(
         }
     }
 
-    override fun update(deltaTime: Double, width: Double, height: Double) {
+    override fun update(deltaTime: Double, width: Int, height: Int) {
         super.update(deltaTime, width, height)
         roadLength += sqrt(speed.sumPow2()) * deltaTime
     }
